Extract out the base toString, hashCode and getUUID code down to an abstract base class
